% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ClassB.R
\docType{class}
\name{B-class}
\alias{B-class}
\title{S4 Class "B"}
\description{
Class "B"
}
\details{
Example class
}
\section{Slots}{

\describe{
\item{\code{xcoords}}{Object of class \code{"list"}}

\item{\code{ycoords}}{Object of class \code{"list"}}
}}
\keyword{classes}

